refactor: split apis
