+++ src/components/rank/Rank.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect } from 'react';
 import Background from '../background';
 import RankImg from '../../assets/images/Rank/rank.png';
 import { RankingType } from '../../types/RankingType';
-import { getPost } from '../apis/api/post';
+import { getPost } from '../../apis/api/post';
 
 const Rank = () => {
   const [ranking, setRanking] = useState<RankingType[]>([]);
